Add casts to enforce uint32_t usage on x86_64
git-svn-id: https://svn.berlios.de/svnroot/repos/escript/branches/0.6.5.experimental@446 5693b1fc-3b75-4070-9b6b-3ce3692a40d5

+++ EScript/Objects/Collections/Array.cpp
@@ -519,7 +519,7 @@ Array::ArrayIterator::~ArrayIterator() {
 
 //! ---|> [Iterator]
 Object * Array::ArrayIterator::key() {
-	return end() ? nullptr : EScript::create(index);
+	return end() ? nullptr : EScript::create(static_cast<uint32_t>(index));
 }
 
 //! ---|> [Iterator]
